Clean all the changes

+++ src/main/java/ballistix/api/capability/CapabilitySiloRegistry.java
@@ -12,11 +12,9 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.NbtUtils;
-import net.minecraftforge.common.ForgeHooks;
 import net.minecraftforge.common.capabilities.Capability;
 import net.minecraftforge.common.capabilities.ICapabilitySerializable;
 import net.minecraftforge.common.util.LazyOptional;
-import net.minecraftforge.server.ServerLifecycleHooks;
 
 public class CapabilitySiloRegistry implements ICapabilitySerializable<CompoundTag> {
 
@@ -53,7 +51,7 @@ public class CapabilitySiloRegistry implements ICapabilitySerializable<CompoundT
 
 			}
 
-			tag.put("frequencytag" + (int) key, posTag);
+			tag.put("frequencytag" + key, posTag);
 
 		});
